Upload of my patterned floor program
This was made for a school project, and is the most complex LUA program I have made to date. Though it is much less complicated than some of my projects in my college CS classes, this was entirely theorized, designed, and coded by myself, instead of having documentation, instructions, or a library file to guide me.